pyproject.toml
[project]
name = "a2a-chat-demo"
version = "0.1.0"
description = "SK + Pydantic AI A2A Chat Demo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "semantic-kernel>=1.30.0",
    "a2a-sdk>=0.2.4",
    "httpx[http2]>=0.28.1",
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
    "fastapi",
    "python-multipart",
]

